Fix crash related to moving WebSockets
Be careful in the capture lists of lambda expressions!
WebSocket was capturing "this" and using it in a delegate
given to the transport connection object.  If the WebSocket
was then moved, the original shell of the WebSocket would
still be used by the delegates.

Fix this in two parts:
1. Capture the impl, not this.  Then the captures of the delegates
   don't change even if the WebSocket is moved.
2. Clear delegates in WebSocket destructor, to prevent them from
   being called in the first place once the WebSocket is destroyed.

Remember to check in the destructor to make sure not to look into
the impl if there is none (e.g. WebSocket was moved).