WebSockets: prevent race condition
During the background processing of a network connection, strong
references back to WebSocket may be obtained in the WebSocket's delegates.
This becomes a problem if the WebSocket is destroyed while the connection
is still being processed.

Prevent a race between destruction of WebSocket and its connection.
In WebSocket's destructor, explicitly break the connection abruptly,
if it's not already broken, to ensure the connection's background
processing is completed, and no strong references back to WebSocket
may be held in any delegates.