Fix buggy usage of async-mutex inside Stream::poll_next
Getting a temporary `MutexGuard` and calling `poll` on it only
works if it's ready.  If not, the task will never be woken
up even if the mutex is unlocked, because we drop the `MutexGuard`
first.

Avoid using a mutex in the first place by using an unbounded channel
instead.  It turns out this is much cleaner since we can leverage
the channel receiver's `poll_next_unpin` method to implement
`Stream::poll_next` completely.