Performance improvement
Only allocate twice the memory as was actually read each loop,
rather than doubling the amount allocated regardless of how
much memory is actually used.

This works well if the transport layer is capped on the size
of the data it reads, as then we won't be wasting memory we
can't use anyway.