Add close timeout
Add third future that stops the worker if a close
isn't received within five seconds after
a close is sent.

Add a test server which immediately sends a close
frame once a WebSocket is opened.

Add a test client which does not send a close
in reply to receiving a close.